change the rule of heart

+++ GameSDLCppDemo/BaseObject.h
@@ -11,7 +11,7 @@ public:
 	~BaseObject();
 	void Show(SDL_Surface* des);
 	bool LoadImg(const char* file_name);
-	void SetRect(const int& x, const int& y) {rect_.x = x, rect_.y = y;}
+	void SetRect(const double& x, const double& y) {rect_.x = x, rect_.y = y;}
 	SDL_Rect GetRect() const {return rect_;}
 	SDL_Surface* GetObject() {return p_object_;}
 protected:
